[Bronze II] Title: 시험 감독, Time: 520 ms, Memory: 5928 KB -BaekjoonHub

+++ 백준/Bronze/13458. 시험 감독/시험 감독.cc
@@ -4,8 +4,8 @@
 using namespace std;
 
 #define ll long long
-ll n, b, c;
-ll arr[1000020];
+int n, b, c;
+int arr[1000020];
 
 int main()
 {
@@ -16,14 +16,14 @@ int main()
     sort(arr, arr+n);
     
     ll cnt = 0;
-    for(ll i = 0; i < n; i++){
-        ll v= arr[i];
+    for(int i = 0; i < n; i++){
+        int v= arr[i];
         // 총감독관 1명
         v -= b;
         cnt++;
         // 부감독관 n
         if(v <= 0) continue;
-        ll delta = v % c == 0 ? 0 : 1;
+        int delta = v % c == 0 ? 0 : 1;
         cnt += v / c + delta;
     }
